requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
python-decouple
dj-database-url
psycopg2-binary
gunicorn
channels
daphne
whitenoise
